Move create_primary_key_if_not_exists to helpers
